fix: rename SchemaCardRenderer to GenuiRenderer
