+++ packages/frameworks/vue/src/renderer/GenuiRenderer.vue
