+++ packages/frameworks/vue/src/renderer/index.ts
@@ -1,3 +1,3 @@
-export { default as GenuiRenderer } from './SchemaCardRenderer.vue'; 
+export { default as GenuiRenderer } from './GenuiRenderer.vue'; 
 export * from './config.js'
 export * from './renderer.types.js';
